pyproject.toml
[project]
name = "bitfinex-margin-balancer"
version = "0.1.0"
description = "Simulate cross-margin behavior on Bitfinex isolated margin"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
    "python-telegram-bot>=20.7",
    "aiosqlite>=0.19.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
